requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# LangChain - AI Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-dashscope>=0.0.1  # DashScope 集成

# LLM SDKs
dashscope>=1.17.0  # 阿里云通义千问

# Vector Database
chromadb>=0.4.0

# Data Validation
pydantic>=2.0.0

# Web Framework and Server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6  # For form data handling in FastAPI

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0  # For clustering and ML features
jieba>=0.42.0  # Chinese text segmentation for topic analysis

# Visualization
matplotlib>=3.7.0  # For charts and graphs
wordcloud>=1.9.0  # For word cloud generation
pillow>=10.0.0  # Image processing for visualizations

# Document Processing
pypdf>=3.0.0  # PDF processing for PyPDFLoader
unstructured>=0.10.0  # Additional document loaders

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0  # For async test support
black>=23.0.0
flake8>=6.1.0

# Optional: Additional utilities
httpx>=0.24.0  # For HTTP requests
tenacity>=8.2.0  # For retry logic
schedule>=1.2.0  # For scheduled tasks
